+++ grunt/compress.js
@@ -21,15 +21,20 @@ module.exports = {
 
       '!bower.json',
       '!bower_components/**',
+      '!views/**/bower.json',
+      '!views/**/composer.json',
+      '!views/**/README*.md',
+      '!views/**/CHANGES*.md',
 
       '!package.json',
       '!node_modules/**',
 
-      '*.phar',
+      '!*.phar',
       '!composer.json',
       '!composer.lock',
       '!vendor/**',
 
+      '!Rakefile',
       '!Gemfile',
       '!Gemfile.lock',
 
@@ -38,12 +43,9 @@ module.exports = {
       '!paths.json',
       '!grunt/**',
 
+      '!RELEASE-LIST.md',
       '!pkg/**',
-
       '!tests/**',
-
-      '!docs/**',
-
       '!bin/**'
 
     ]